commit N°6
Añadí otro proyecto RestApp, una forma de implementar Rest pero con Dao

+++ wsCrud/src/main/java/com/ws/dao/InterfaceProductsDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ws.entity.Products;
 
-public interface InterfaceProductsDao extends CrudRepository<Products, Integer> {
+public interface InterfaceProductsDao extends CrudRepository<Products, Long> {
 
 }
